# openapi.yaml
openapi: 3.0.0
info:
  title: SportsDeck API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/users/{id}/activity:
    get:
      summary: Get a user's daily activity counts over a specified time range
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clxuser789
        - in: query
          name: range
          schema:
            type: string
            enum:
              - 7d
              - 30d
              - 90d
              - 1y
          example: 30d
      responses:
        '200':
          description: Daily activity data including posts, replies, and totals
        '404':
          description: User not found
        '500':
          description: Internal server error
...